Final Version of Dicoding OpenMusic Project

+++ javascript/open-music-api/src/services/postgres/PlaylistsService.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-underscore-dangle */
 const { Pool } = require('pg');
 const { nanoid } = require('nanoid');
 const InvariantError = require('../../exceptions/InvariantError');
@@ -7,9 +6,10 @@ const AuthorizationError = require('../../exceptions/AuthorizationError');
 const { mapDBToModel } = require('../../utils/playlists');
 
 class PlaylistsService {
-  constructor(collaborationsService) {
-    this._pool = new Pool();
-    this._collaborationsService = collaborationsService;
+  constructor(collaborationsService, cacheService) {
+    this.pool = new Pool();
+    this.collaborationsService = collaborationsService;
+    this.cacheService = cacheService;
   }
 
   async addPlaylist({ name, owner }) {
@@ -20,22 +20,38 @@ class PlaylistsService {
       values: [id, name, owner],
     };
 
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
 
     if (!result.rowCount) {
       throw new InvariantError('Playlist gagal ditambahkan');
     }
 
+    await this.cacheService.delete(`playlists:${owner}`);
+
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON playlists.id = collaborations.playlist_id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows;
+    try {
+      const result = await this.cacheService.get(`playlists:${owner}`);
+      return {
+        playlists: JSON.parse(result),
+        cached: true,
+      };
+    } catch (error) {
+      const query = {
+        text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner LEFT JOIN collaborations ON playlists.id = collaborations.playlist_id WHERE playlists.owner = $1 OR collaborations.user_id = $1',
+        values: [owner],
+      };
+      const result = await this.pool.query(query);
+
+      await this.cacheService.set(`playlists:${owner}`, JSON.stringify(result.rows), 1800);
+
+      return {
+        playlists: result.rows,
+        cached: false,
+      };
+    }
   }
 
   async getPlaylistById(id) {
@@ -43,7 +59,7 @@ class PlaylistsService {
       text: 'SELECT playlists.id, playlists.name, users.username FROM playlists LEFT JOIN users ON users.id = playlists.owner WHERE playlists.id = $1',
       values: [id],
     };
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
 
     if (!result.rowCount) {
       throw new NotFoundError('Playlist tidak ditemukan');
@@ -54,15 +70,18 @@ class PlaylistsService {
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
+    const { owner } = result.rows[0];
 
     if (!result.rowCount) {
       throw new NotFoundError('Playlist gagal dihapus. Id tidak ditemukan');
     }
+
+    this.cacheService.delete(`playlists:${owner}`);
   }
 
   async verifyPlaylistOwner(id, owner) {
@@ -71,7 +90,7 @@ class PlaylistsService {
       values: [id],
     };
 
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
 
     if (!result.rowCount) {
       throw new NotFoundError('Playlist tidak ditemukan');
@@ -92,7 +111,7 @@ class PlaylistsService {
         throw error;
       }
       try {
-        await this._collaborationsService.verifyCollaborator(playlistId, userId);
+        await this.collaborationsService.verifyCollaborator(playlistId, userId);
       } catch {
         throw error;
       }
@@ -106,20 +125,36 @@ class PlaylistsService {
       values: [id, playlistId, songId],
     };
 
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
 
     if (!result.rowCount) {
       throw new InvariantError('Lagu gagal ditambahkan ke playlist');
     }
+
+    this.cacheService.delete(`playlistsongs:${playlistId}`);
   }
 
   async getSongsFromPlaylist(playlistId) {
-    const query = {
-      text: 'SELECT songs.id, songs.title, songs.performer FROM songs LEFT JOIN playlist_songs ON songs.id = playlist_songs.song_id WHERE playlist_songs.playlist_id = $1',
-      values: [playlistId],
-    };
-    const result = await this._pool.query(query);
-    return result.rows;
+    try {
+      const result = await this.cacheService.get(`playlistsongs:${playlistId}`);
+      return {
+        songs: JSON.parse(result),
+        cached: true,
+      };
+    } catch (error) {
+      const query = {
+        text: 'SELECT songs.id, songs.title, songs.performer FROM songs LEFT JOIN playlist_songs ON songs.id = playlist_songs.song_id WHERE playlist_songs.playlist_id = $1',
+        values: [playlistId],
+      };
+      const result = await this.pool.query(query);
+
+      await this.cacheService.set(`playlistsongs:${playlistId}`, JSON.stringify(result.rows), 600);
+
+      return {
+        songs: result.rows,
+        cached: false,
+      };
+    }
   }
 
   async deleteSongFromPlaylist(playlistId, songId) {
@@ -128,11 +163,13 @@ class PlaylistsService {
       values: [playlistId, songId],
     };
 
-    const result = await this._pool.query(query);
+    const result = await this.pool.query(query);
 
     if (!result.rowCount) {
       throw new InvariantError('Lagu gagal dihapus dari playlist. Id tidak ditemukan');
     }
+
+    this.cacheService.delete(`playlistsongs:${playlistId}`);
   }
 }
